+++ src/easy_debug.h
@@ -30,8 +30,10 @@
  *
  **/
 
+#ifndef EASYMAC2STR
 #define EASYMAC2STR(a) (a)[0], (a)[1], (a)[2], (a)[3], (a)[4], (a)[5]
 #define EASYMACSTR "%02X:%02X:%02X:%02X:%02X:%02X"
+#endif
 
 // Define the log levels
 #define LOG_NONE 0